[openfda] Beginnings of more robust query builder.
Implements a simple mapping between prop names and their corresponding
openFDA predicate names. Unfortunately, there are more props than just
the predicates, so the predicate values need to reside in their own
object. This doesn't work with valueLink binding (though it is in the
works; see https://github.com/facebook/react/issues/744).

I think the solution here is to create a new component for the inputs
instead of the simple bootstrap-react one.